+++ samples/observer-demo/src/main/kotlin/net/corda/observerdemo/flow/RegistryException.kt
@@ -1,0 +1,3 @@
+package net.corda.observerdemo.flow
+
+class RegistryException(val error: RegistryObserverFlow.Error) : Exception("Error response from trade finance registry - ${error}")